🔧 FIX: Arduino serial communication and compilation issues
## 🎯 Key Improvements
- **Arduino Protocol Compatibility**: Fixed command format from uppercase to lowercase (M -> m) for Arduino compatibility
- **Native Windows API Migration**: Migrated from serial library to native Windows API for better performance and reliability
- **Auto-reconnection System**: Added intelligent connection health monitoring and automatic reconnection
- **CUDA Compilation**: Enabled previously disabled CUDA files in project configuration

## 🛠️ Technical Fixes
- **Compilation Warnings**: Resolved C4005 (macro redefinition), C4838 (narrowing conversion), C4819 (encoding), C2065 (undeclared identifier)
- **Serial Communication**: Implemented synchronous I/O with improved error handling and retry logic
- **Memory Management**: Enhanced connection state management with thread-safe operations
- **Project Configuration**: Updated vcxproj to include CUDA files and UTF-8 compiler flags

## 🚀 Performance Enhancements
- **Reduced Latency**: Eliminated unnecessary TIMER commands and debug logging
- **Improved Reliability**: Added connection validation and graceful degradation
- **Optimized Timeouts**: Configured appropriate timeouts for Arduino communication

## ✅ Validation
- Mouse movement functionality restored and working
- Recoil compensation system operational
- All compilation warnings resolved
- Arduino protocol compatibility confirmed

Fixes critical Arduino communication issues and eliminates all compilation warnings while maintaining functionality.

+++ needaimbot/config/config_validator.h
@@ -44,7 +44,7 @@ public:
         // Recoil settings
         config.easynorecoilstrength = std::clamp(config.easynorecoilstrength, 0.0f, 10.0f);
         config.norecoil_step = std::clamp(config.norecoil_step, 0.1f, 5.0f);
-        config.norecoil_ms = (std::max)(1, (std::min)(static_cast<int>(config.norecoil_ms), 100));
+        config.norecoil_ms = static_cast<float>((std::max)(1, (std::min)(static_cast<int>(config.norecoil_ms), 100)));
         config.easynorecoil_start_delay_ms = (std::max)(0, config.easynorecoil_start_delay_ms);
         config.easynorecoil_end_delay_ms = (std::max)(0, config.easynorecoil_end_delay_ms);
 
@@ -78,7 +78,7 @@ public:
             UtilConstants::MAX_OVERLAY_OPACITY
         );
         config.overlay_ui_scale = std::clamp(config.overlay_ui_scale, 0.5f, 3.0f);
-        config.window_size = (std::max)(0.1f, (std::min)(static_cast<float>(config.window_size), 3.0f));
+        config.window_size = static_cast<int>((std::max)(0.1f, (std::min)(static_cast<float>(config.window_size), 3.0f)));
 
         // Screenshot delay
         config.screenshot_delay = (std::max)(0, config.screenshot_delay);
